pyproject.toml
[build-system]
requires = ["setuptools==75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "news-for-dumb"
description = "AI summerized news."
# dynamic = ["version"]
version = "0.0.1"
readme = "README.md"

authors = [
    {name = "Gabriel Saganić", email = "gabriel.saganic@gmail.com"}
]
maintainers = [
    {name = "Gabriel Saganić", email = "gabriel.saganic@gmail.com"}
]
requires-python = ">=3.10"
dependencies=[
    "django==5.1.*",
    "newsapi-python==0.2.7",
    "beautifulsoup4==4.13.3",
    "pytz==2025.1",
    "boto3==1.37.18",
    "python-dotenv==1.0.1",
    "djangorestframework==3.15.2",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "openai==1.70.0",
#    "torch==2.6.0",
#    "transformers==4.50.3",
#    "sentencepiece==0.2.0",
]
keywords = ["news", "AI", "summerized"]

[project.optional-dependencies]
dev = [
  "black==25.1.0",
  "isort==6.0.1"
]

[project.scripts]
"manage.py" = "myproject:django_manage"

[tool.isort]
profile = "black"
skip = ["migrations", "settings.py", "admin.py"]

[tool.black]
line-length = 88
exclude = '''
(
    migrations/.*
    | settings.py$
    | admin.py$
)
'''